+++ packages/memory-store/src/memory-store.js
@@ -17,6 +17,7 @@ import { create as createAction } from "./create.js";
 import { deleteAction } from "./delete.js";
 import { update as updateAction } from "./update.js";
 import { merge } from "./merge.js";
+import { createIdGenerator } from "./lib/store-helpers.js";
 
 /**
  * @typedef {Object} NormalResourceTree
@@ -40,6 +41,7 @@ import { merge } from "./merge.js";
  * @property {Ajv} [validator] - AJV validator instance for resource validation
  * @property {MemoryStore} [store] - The memory store instance
  * @property {import('@spectragraph/core').Graph} storeGraph - The graph data structure containing all resources
+ * @property {function(string): string|number} idGenerator - Function that generates IDs for resources
  */
 
 /**
@@ -69,6 +71,7 @@ export function createMemoryStore(schema, config = {}) {
 	ensureValidSchema(schema, { validator });
 
 	let storeGraph = mergeGraphsDeep(createEmptyGraph(schema), initialData);
+	const idGenerator = createIdGenerator(schema, initialData);
 
 	const runQuery = (query) => {
 		const normalQuery = normalizeQuery(schema, query, {
@@ -85,7 +88,7 @@ export function createMemoryStore(schema, config = {}) {
 	// WARNING: MUTATES storeGraph
 	const create = storeMutation(schema, "create", (normalResource) => {
 		ensureValidCreateResource(schema, normalResource, validator);
-		return createAction(normalResource, { schema, storeGraph });
+		return createAction(normalResource, { schema, storeGraph, idGenerator });
 	});
 
 	// WARNING: MUTATES storeGraph
@@ -128,7 +131,12 @@ export function createMemoryStore(schema, config = {}) {
 			return Promise.resolve(runQuery(query));
 		},
 		async merge(resourceTreeOrTrees) {
-			return merge(resourceTreeOrTrees, { schema, storeGraph, validator });
+			return merge(resourceTreeOrTrees, {
+				schema,
+				storeGraph,
+				validator,
+				idGenerator,
+			});
 		},
 	};
 }